+++ processor/src/test/resources/bad_input/com/example/reserved_names/FieldPrefixReservedNameModel.java
@@ -1,0 +1,9 @@
+package com.example.reserved_names;
+
+import arez.annotations.ArezComponent;
+
+@ArezComponent( allowEmpty = true )
+public abstract class FieldPrefixReservedNameModel
+{
+  abstract void $$arez$$_hello( String v );
+}